clean up

+++ syscall.c
@@ -1,4 +1,4 @@
-typedef unsigned long size_t;
+typedef unsigned long long size_t;
 
 size_t __syscall(size_t nr, size_t a0, size_t a1, size_t a2, size_t a3, size_t a4, size_t a5) {
     size_t ret;
